requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
djangorestframework==3.15.2
frozenlist==1.4.1
idna==3.8
multidict==6.0.5
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
twilio==9.2.3
tzdata==2024.1
<<<<<<< HEAD
django-cors-headers==4.4.0
=======
urllib3==2.2.2
yarl==1.9.4
>>>>>>> c475940c9052cbc3590b403851ba38afd5f49594
